refactor: server, config and userscript plugin
