+++ src/banner.ts
@@ -2,19 +2,17 @@ import type { HeaderConfig } from './types.js'
 
 export class Banner {
   private header: string[] = []
-  private configKeys: string[]
   private maxKeyLength: number
 
   constructor(private readonly config: HeaderConfig) {
     this.downloadMeta()
-    this.configKeys = Object.keys(this.config)
     this.maxKeyLength =
-      Math.max(...this.configKeys.map((key) => key.length)) + 1
+      Math.max(...Object.keys(this.config).map((key) => key.length)) + 1
   }
 
   private downloadMeta(): void {
-    const { name, homepage, updateURL, downloadURL } = this.config
-    if (homepage && !updateURL && !downloadURL) {
+    const { name, homepage } = this.config
+    if (homepage) {
       this.config.updateURL = new URL(`${name}.meta.js`, homepage).href
       this.config.downloadURL = new URL(`${name}.user.js`, homepage).href
     }
@@ -24,10 +22,13 @@ export class Banner {
     return ' '.repeat(this.maxKeyLength - str.length)
   }
 
-  private addMetadata(key: string, value: string | string[] | number | boolean): void {
+  private addMetadata(
+    key: string,
+    value: string | string[] | number | boolean
+  ): void {
     value = Array.isArray(value) ? value.join(' ') : value === true ? '' : value
-    const spaces = this.addSpaces(key)
-    this.header.push(`// @${key}${spaces}${value}`)
+
+    this.header.push(`// @${key}${this.addSpaces(key)}${value}`)
   }
 
   generate(): string {
