+++ src/constants.ts
@@ -1,6 +1,5 @@
 import type { Grants } from './types.js'
 
-export const vitePluginName = 'vite-userscript-plugin'
 export const styleTemplate = 'console.warn("__STYLE__")'
 export const regexpScripts = new RegExp(/\.(t|j)sx?$/)
 export const regexpStyles = new RegExp(/\.(s[ac]|c|le)ss$/)
